Move src/components/literature/ to src/apps/mydb/elements/literature/ and move the subtab component DetailsTabLiteratures into the folder
