+++ app/packs/src/apps/mydb/elements/details/literature/CitationType.js
