+++ app/packs/src/apps/mydb/elements/details/researchPlans/ResearchPlanDetails.js
@@ -8,7 +8,7 @@ import UIActions from 'src/stores/alt/actions/UIActions';
 import ElementActions from 'src/stores/alt/actions/ElementActions';
 import DetailActions from 'src/stores/alt/actions/DetailActions';
 import ResearchPlansFetcher from 'src/fetchers/ResearchPlansFetcher';
-import ResearchPlansLiteratures from 'src/apps/mydb/elements/details/DetailsTabLiteratures';
+import ResearchPlansLiteratures from 'src/apps/mydb/elements/details/literature/DetailsTabLiteratures';
 import ResearchPlanWellplates from 'src/apps/mydb/elements/details/researchPlans/wellplatesTab/ResearchPlanWellplates';
 import ResearchPlanMetadata from 'src/apps/mydb/elements/details/researchPlans/ResearchPlanMetadata';
 import Attachment from 'src/models/Attachment';
